+++ ui/preferences/src/main/java/de/danoeh/antennapod/ui/preferences/screen/bugreport/BugReportFragment.java
@@ -85,9 +85,9 @@ public class BugReportFragment extends AnimatedFragment {
             }
         });
         viewBinding.openForumButton.setOnClickListener(v ->
-                IntentUtils.openInBrowser(requireContext(), "https://forum.antennapod.org/search"));
+                IntentUtils.openInBrowser(requireContext(), "https://github.com/mpdavison/AntennaPod/issues"));
         viewBinding.openGithubButton.setOnClickListener(v ->
-                IntentUtils.openInBrowser(requireContext(), "https://github.com/AntennaPod/AntennaPod/issues"));
+                IntentUtils.openInBrowser(requireContext(), "https://github.com/mpdavison/AntennaPod/issues"));
         viewBinding.attribAppVersionLabel.setOnClickListener(v ->
                 ClipboardUtils.copyText((TextView) v, R.string.report_bug_attrib_app_version));
         viewBinding.attribAndroidVersionLabel.setOnClickListener(v ->